Update_Workflow_0509
Update_Workflow_0509

+++ AIops_BackEnd/src/main/java/com/aiops_web/controller/OriginalDataController.java
@@ -32,7 +32,7 @@ public class OriginalDataController {
 
     @GetMapping("/relativeRange")
     public ResponseStd<List<OriginalData>> getRelativeRange(int batchId, long beginId, long endId) {
-        return new ResponseStd<>(originalDataService.getRelativeRange(batchId, beginId, endId));
+        return new ResponseStd<>(originalDataService.getRelativeRange(batchId, (int)beginId, (int)endId));
     }
 
     @DeleteMapping("/range")
